Moves fakePassingExchange to test_helper. Fixes some order side implementations for trades. Botches a new error type
